Add elpa package (evil)
The alternative would be to have some logic in the `init.el` that
installs packages when it detects that they have not been installed yet.
This has the advantage that less needs to be in this directory, but at
the cost of reproducibility. I don't mind storing all the packages, but
will probably put all of this emacs stuff in a different git repo and
include it as a submodule. I will probably do something similar for
neovim as well eventually.

For now I only added evil-mode. Will add more eventually.